p16-comments-AcceptModal-logic

+++ src/Components/Comments.js
@@ -3,12 +3,15 @@ import Errorbox from './Errorbox'
 import DetailsModal from './DetailsModal'
 import DeleteModal from './DeleteModal'
 import EditModal from './EditModal'
+import {BsCheckCircle} from 'react-icons/bs'
 export default function Comments() {
   const [allComments, setAllComments] = useState([])
   const [isSowDetailsModal, setIsSowDetailsModal] = useState(false)
   const [isSowDeleteModal, setIsSowDeleteModal] = useState(false)
   const [isSowEditModal, setIsSowEditModal] = useState(false)
+  const [isShowAcceptModal, setIsShowAcceptModal] = useState(false)
   const [mainCommentID, setMainCommentID] = useState(null)
+  const [mainCommentIDAccept, setMainCommentIDAccept] = useState(null)
   const [mainCommentBody, setMainCommentBody] = useState('')
   const [mainCommentInfo, setMainCommentInfo] = useState({})
   
@@ -49,17 +52,34 @@ export default function Comments() {
   }, [])
 
   const submitAction = () => {
-    fetch(`http://localhost:8000/api/comments/${mainCommentID}`, {
-      method: 'DELETE'
-    })
-      .then(res => res.json())
-      .then(data => {
-        getAllComments()
-        setIsSowDeleteModal(false)
+    if(mainCommentID){
+      fetch(`http://localhost:8000/api/comments/${mainCommentID}`, {
+        method: 'DELETE'
       })
-  }
-  const cancelAction = () => {
-    setIsSowDeleteModal(false)
+        .then(res => res.json())
+        .then(data => {
+          getAllComments()
+          setIsSowDeleteModal(false)
+          setMainCommentID(null)
+        })
+      }
+      
+      if(mainCommentIDAccept){
+        fetch(`http://localhost:8000/api/comments/accept/${mainCommentIDAccept}`,{
+          method: 'POST'
+        })
+        .then(res=>res.json())
+        .then(data =>{
+          console.log(data);
+          setIsShowAcceptModal(false)
+          getAllComments()
+        })
+    }
+
+    }
+    const cancelAction = () => {
+      setIsSowDeleteModal(false)
+      setIsShowAcceptModal(false)
 
   }
 
@@ -95,7 +115,7 @@ export default function Comments() {
                       }} className='bg-violet-900 rounded-md px-4 py-1 text-white'>دیدن متن</button></td>
                       <td>{comment.date}</td>
                       <td>{comment.hour}</td>
-                      <td className=' h-20 flex justify-center gap-2 text-white items-center'>
+                      <td className=' h-20 flex justify-center ms-5 gap-2 text-white items-center'>
                         <button onClick={() => {
                           setMainCommentID(comment.id)
                           setIsSowDeleteModal(true)
@@ -106,7 +126,19 @@ export default function Comments() {
                           setMainCommentBody(comment.body)
                         }} className='bg-violet-900 rounded-md px-4 py-1'>ویرایش</button>
                         <button className='bg-violet-900 rounded-md px-4 py-1'>پاسخ</button>
-                        <button className='bg-violet-900 rounded-md px-4 py-1'>تایید</button>
+                        {!comment.isAccept && (
+                          <button onClick={()=>{
+                            setIsShowAcceptModal(true)
+                            setMainCommentIDAccept(comment.id)
+                          }} className='bg-violet-900 rounded-md px-4 py-1'>تایید</button>
+                        )}
+                        {
+                          comment.isAccept && (
+                            <div className='text-green-700 flex justify-center px-4 py-1 text-4xl'>
+                              <BsCheckCircle></BsCheckCircle>
+                            </div>
+                          )
+                        }
                       </td>
                     </tr>
                   ))
@@ -126,7 +158,10 @@ export default function Comments() {
       </DetailsModal>
 
 
-      <DeleteModal isSowDeleteModal={isSowDeleteModal} submitAction={submitAction} cancelAction={cancelAction}></DeleteModal>
+      <DeleteModal isShowAcceptModal={isShowAcceptModal} isSowDeleteModal={isSowDeleteModal} submitAction={submitAction} cancelAction={cancelAction} title={isSowDeleteModal ? `آیا از   حذف کامنت مطمعن هستید؟` : isShowAcceptModal ? `آیا کامنت مورد تایید است؟` : ''} ></DeleteModal>
+
+
+      
 
       <EditModal onSubmit={onSubmit} isSowEditModal={isSowEditModal} onClose={onClose}>
       <textarea onChange={(e)=>setMainCommentBody(e.target.value)} value={mainCommentBody} cols="30" rows="10">
